Teensy update. Not pushing to package yet.
Just a few small patches, not big or important enough to call for their own package update (Plus I have stuff to change in package.json before I'd push 0.1.3 or higher).
I've just marked as 0.1.2-beta.0 to easier see that this is not the prod version of the package, and this is still in dev.